mambo: Add util function for breaking on console output
Mambo can execute a trigger when certain output appears on the console.
You can run any tcl function when the trigger fires, but the simplest
thing to use it for is stopping the simulation.

Add a helper to do that, break_on_console(), and a matching function to
clear the trigger, clear_console_break().

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>