phb4: Enforce root complex config space size of 2048
The root complex config space size on PHB4 is 2048. This patch sets
that size and enforces it when trying to read/write the config space
in the root complex.

Without this someone reading the config space via /sysfs in linux will
cause an EEH on the PHB.

If too high, reads returns 1s and writes are silently dropped.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>