core/pci: Rename pci_slot_op poll to run_sm
This renames the "poll" op to "run_sm" (short for run state machine).

I think this is a better name since the function does a bunch of
things like reseting the slot.  Also it avoids confusion with the
"poll_link" op which does something different (and can even be called
from run_sm).

No functional change.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Acked-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>