mambo: Make xscom claim to be DD 2.0
In the mambo tcl we set the CPU version to DD 2.0, because mambo is not
bug compatible with DD 1.

But in xscom_read_cfam_chipid() we have a hard coded value, to work
around the lack of the f000f register, which claims to be P9 DD 1.0.

This doesn't seem to cause crashes or anything, but at boot we do see:

  [    0.003893084,5] XSCOM: chip 0x0 at 0x1a0000000000 [P9N DD1.0]

So fix it to claim that the xscom is also DD 2.0 to match the CPU.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>