chip: Add support for discovering chips on mambo
Currently the only way for skiboot to discover chips is by looking for xscom
nodes. But on mambo it's currently not possible to create multiple xscom nodes,
which means we can only simulate a single chip system.

However it seems we can fairly cleanly add support for a special mambo chip
node, and use that to instantiate multiple chips.

Add a check in init_chip() that we're not clobbering an already initialised
chip, now that we have two places that initialise chips.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>