hw/fsp: Decode payload initrd on FSP machines
During bootup on FSP systems, the payload kernel and initrd files are
loaded seperately. While the initrd has been loaded, it is still in its
compressed form. The initrd is generally compressed using xz encoding.

As we have plenty of spare resources during the Skiboot initialisation we
can preemptively decode the initrd file before the Linux kernel needs it.
This saves time during the kernel boot as it doesn't need to wait for the
initrd to decompress before continuing.

This patch adds the load_done function to platforms, with a FSP
specific implementation, which loads and configures the boot resources.
This consists of asyncronously loading and then decoding the initramfs.

By using the asyncronous load_done function we reduce the boot time by 2.5
seconds.

Note: This requires the skiboot libxz patches

Signed-off-by: Matt Brown <matthew.brown.dev@gmail.com>