core: Decode payload initrd on openpower
This patch adds the ability for openpower machines to decode the
payloads initramfs before the kernel requires it. Allowing petitboot to
boot faster.

Openpower machines use zImage files for the payload. This contains
the kernel, with the initramfs baked into the file. There is not way to
determine the location of the initramfs this way.

In order to do the asyncronous decoding on openpower, the payload must be
packaged as a Flattened Image Tree (FIT). This is similar to a standard DT,
but contains the kernel and initramfs images. This allows us to locate the
initramfs and decode it before the kernel requires it.

This gives similar performance benefits to the FSP case, saving around 2.7
seconds during the boot process.

To create the FIT file we use the `mkimage` tool. You can find
documentation on FIT files and how to create them on from the u-boot
repository. (Under /doc/uImage.FIT/)

Signed-off-by: Matt Brown <matthew.brown.dev@gmail.com>