phb4: Harden init with bad PHBs
Currently if we read all 1's from the EEH or IRQ capabilities, we end
up train wrecking on some other random code (eg. an assert() in xive).

This hardens the PHB4 code to look for these bad reads and more
gracefully fails the init for that PHB alone.  This allows the rest of
the system to boot and ignore those bad PHBs.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>